Sprint 1 Demo Completed *READ MORE*
I am a big dumb idiot and did all of the previous coding in react instead of react native. Everything is pretty similar except the map library (Leaflet) that was originally used is not compatible with react native and for that reason I had to start everything over from square one. But(!!), after spending my Friday at the computer we're caught back up to where we were lol